% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cyclical_sampler.R
\name{Cyclical_sampler}
\alias{Cyclical_sampler}
\title{Cyclical MCMC sampler (column-wise, data-augmented Gibbs)}
\usage{
Cyclical_sampler(S, n, M, burnin, prior)
}
\arguments{
\item{S}{sample covariance matrix}

\item{n}{sample size}

\item{M}{Total number of MCMC iterations}

\item{burnin}{number of MCMC burnins}

\item{prior}{character, one of \code{"GHS"} or \code{"GHSL"}.}
}
\value{
A list \code{list(omega, lambda_sq, tau_sq)} as described above.
}
\description{
\code{Cyclical_sampler} returns a list with:
\itemize{
\item \code{omega}: array \eqn{p \times p \times M-burnin} of saved precision matrices;
\item \code{lambda_sq}: array \eqn{p \times p \times M-burnin} of local parameters;
\item \code{tau_sq}: numeric vector (\eqn{M-burnin}) of the global parameter.
}
}
